separated file about Configuration
